Rename topology_manager to topology_util
