+++ modi/util/topology_util.py
